// src/model/productmodel.ts
export interface ProductModel {
    handle: string;
    title: string;
    body_html: string;
    vendor: string;
    product_type: string;
    tags: string;
    published: string;
    option1_name: string;
    option1_value: string;
    option2_name: string;
    option2_value: string;
    option3_name: string
    option3_value: string
    variant_sku: string
    variantgrams: string
    variant_inventory_tracker: string
    variant_inventory_qty: string
    variant_inventory_policy: string
    variant_fulfillment_service: string
    variant_price: string
    variant_compare_at_price: string
    variant_requires_shipping: string
    variant_taxable: string
    variant_barcode: string
    image_src: string
    image_position: string
    image_alt_text: string
    gift_card: string
    seo_title: string
    seo_description: string
    google_shopping_google_product_category: string
    google_shopping_gender: string
    google_shopping_agegroup: string
    google_shopping_mpn: string
    google_shopping_adWordsgrouping: string
    google_shopping_adWords_labels: string
    google_shopping_condition: string
    google_shopping_custom_product: string
    google_shopping_custom_label_0: string
    google_shopping_custom_label_1: string
    google_shopping_custom_label_2: string
    google_shopping_custom_label_3: string
    google_shopping_custom_label_4: string
    variant_image: string
    variant_weight_unit: string
    variant_tax_code: string
}